Extract common code for clients
